% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parse_raw_data.R
\name{initiate_canonical_databases}
\alias{initiate_canonical_databases}
\title{initiate_canonical_databases}
\usage{
initiate_canonical_databases(
cosmic_file = "CosmicCLP_MutantExport.tsv",
ccle_file = "CCLE_hybrid_capture1650_hg19_NoCommonSNPs_CDS_2012.05.07.maf",
ref_gen = "GRCH37")
}
\arguments{
\item{cosmic_file}{The path to the cosmic DNA genotype data file. 
Ensure that the right reference genome is used}

\item{ccle_file}{The path to the ccle DNA genotype data file. 
Ensure that the right reference genome is used}

\item{ref_gen}{Reference genome version}
}
\value{
Returns message if parsing process has succeeded
}
\description{
Parses data into r list variable
}
\examples{
initiate_canonical_databases(
cosmic_file = "CosmicCLP_MutantExport.tsv",
ccle_file = "CCLE_hybrid_capture1650_hg19_NoCommonSNPs_CDS_2012.05.07.maf",
ref_gen = "GRCH37")
}

