% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parser_scripts.R
\name{parse_vcf_file}
\alias{parse_vcf_file}
\title{parse_vcf_file}
\usage{
parse_vcf_file( vcf_file_path, n_threads)
}
\arguments{
\item{vcf_file_path}{Path to the vcf file on the operating system}

\item{n_threads}{Specifies number of threads to be used}
}
\value{
Loci-based DNA-mutational fingerprint of the cancer cell line as found in the input VCF file
}
\description{
Parses the vcf file and filters all information except for the start and length of variations/ mutations.
}
