% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Show_database_information.R
\name{show_contained_mutations_for_cl}
\alias{show_contained_mutations_for_cl}
\title{show_contained_mutations_for_cl}
\usage{
show_contained_mutations_for_cl( 
name_cl, 
ref_gen)
}
\arguments{
\item{name_cl}{Name of the cancer cell line sample stored in the database}

\item{ref_gen}{Reference genome version}
}
\value{
R table which contains all mutations associated with the defined cancer cell line and reference genome
}
\description{
Show all mutations present in the database for a selected cancer cell line and reference Genome
}
\examples{
SK_OV_3_CELLMINER_mutations = show_contained_mutations_for_cl(
name_cl = "SK_OV_3_CELLMINER_mutations",
ref_gen = "GRCH37")
}

