% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parser_scripts.R
\name{split_add}
\alias{split_add}
\title{split_add}
\usage{
split_add(vcf_matrix_row)
}
\arguments{
\item{vcf_matrix_row}{row of the vcf file}
}
\value{
Transformed entry of vcf file, reduced to start and length
}
\description{
split_add
}
