% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parse_VCF_data.R
\name{parse_vcf_file}
\alias{parse_vcf_file}
\title{parse_vcf_file}
\usage{
parse_vcf_file( vcf_file_path )
}
\arguments{
\item{vcf_file_path}{Path to the vcf file on the operating system}
}
\value{
Loci-based DNA-mutational fingerprint of the cancer cell line as found in the input VCF file
}
\description{
Parses the vcf file and filters all information except for the start and length of variations/ mutations.
}

